pyproject.toml
[project]
name = "populis-protocol"
version = "0.1.0"
requires-python = ">= 3.10"
dependencies = [
  "chia-blockchain>=2.3.0",
  "blspy",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.7",
]

[tool.pytest.ini_options]
minversion = "8.1.1"
pythonpath = ["."]
testpaths = ["tests"]
log_cli = 1
log_cli_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["populis_puzzles*"]
